requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
pillow
streamlit-camera-input-live
